requirements.txt
Django==2.1.7
django-ckeditor==5.6.1
django-cleanup==3.2.0
django-crispy-forms==1.7.2
django-js-asset==1.2.2
django-recaptcha==2.0.2
Pillow==5.4.1
pytz==2018.9